requirements.txt
astroid==2.2.5
autopep8==1.4.4
certifi==2019.3.9
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
Django==2.2
django-allauth==0.39.1
django-cors-headers==2.5.2
django-rest-auth==0.9.5
djangorestframework==3.9.2
djangorestframework-jwt==1.11.0
gunicorn==19.9.0
idna==2.8
isort==4.3.17
itypes==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.3.1
Markdown==3.1
MarkupSafe==1.1.1
mccabe==0.6.1
oauthlib==3.0.1
pep8==1.7.1
Pillow==6.0.0
pkg-resources==0.0.0
pycodestyle==2.5.0
Pygments==2.3.1
PyJWT==1.7.1
pylint==2.3.1
python3-openid==3.1.0
pytz==2019.1
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
sqlparse==0.3.0
typed-ast==1.3.4
uritemplate==3.0.0
urllib3==1.24.2
wrapt==1.11.1
